producer/requirements.txt
psycopg[binary]
kafka-python
python-dotenv
